Extract metrics out into module
